+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Releases
 ========
 
+Unreleased
+----------
+
+* `#547 <https://github.com/pytest-dev/pytest-mock/issues/547>`_: Added ``SpyType`` for annotating ``mocker.spy`` results.
+
 3.15.1
 ------